Lightwave 3D Exporter: -Added some Status Bar messages.
DevWork:
-Built & separated Visual Studio 2008 & 2010 versions of WMV's project files.
-StormLib is now included in the VS Project files and will be compiled prior to WMV being built.
-Recompiled & renamed all the .lib files.
-Removed 2 unnecessary lib files from our build of StormLib.
-Minor code fixes.

Wiki:
-Updated Windows Compilation page.

+++ src/UserSkins.cpp
@@ -50,7 +50,7 @@ void UserSkins::LoadFile(const wxString &filename)
 			return;
 		}
 
-		int numGroups = atoi(line.c_str());
+		size_t numGroups = atoi(line.c_str());
 		if (numGroups < 0) {
 			wxLogMessage(wxT("Error - UserSkins: negativ number of groups specified in line %d"), lineNr);
 			return;
